Moved test files next to the actual files to make things easier
